update: Cleaning the code
