--- src/components/forms/AddParkForm.tsx
@@ -1,118 +1,0 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { useRouter } from "next/navigation";
-import { toast } from "sonner";
-
-const parkFormSchema = z.object({
-  name: z.string().min(2, {
-    message: "Name must be at least 2 characters.",
-  }),
-  location: z.string().min(2, {
-    message: "Location must be at least 2 characters.",
-  }),
-  description: z.string().min(10, {
-    message: "Description must be at least 10 characters.",
-  }),
-});
-
-type ParkFormValues = z.infer<typeof parkFormSchema>;
-
-export default function AddParkForm() {
-  const router = useRouter();
-  const [createPark, { isLoading }] = useCreateParkMutation();
-
-  const form = useForm<ParkFormValues>({
-    resolver: zodResolver(parkFormSchema),
-    defaultValues: {
-      name: "",
-      location: "",
-      description: "",
-    },
-  });
-
-  const onSubmit = async (data: ParkFormValues) => {
-    try {
-      await createPark(data).unwrap();
-      toast.success("Park created successfully");
-      router.push("/admin/parks");
-    } catch (error: any) {
-      toast.error(error.data?.message || "Failed to create park");
-    }
-  };
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-        <FormField
-          control={form.control}
-          name="name"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Park Name</FormLabel>
-              <FormControl>
-                <Input placeholder="Nyungwe National Park" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="location"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Location</FormLabel>
-              <FormControl>
-                <Input placeholder="Southwest Rwanda" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="description"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Description</FormLabel>
-              <FormControl>
-                <Textarea
-                  placeholder="Describe the park's features, wildlife, and attractions"
-                  className="min-h-[120px]"
-                  {...field}
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <div className="flex justify-end space-x-4">
-          <Button
-            type="button"
-            variant="outline"
-            onClick={() => router.push("/admin/parks")}
-          >
-            Cancel
-          </Button>
-          <Button type="submit" disabled={isLoading}>
-            {isLoading ? "Creating..." : "Create Park"}
-          </Button>
-        </div>
-      </form>
-    </Form>
-  );
-}
